# src/NovaLab.Api.Twitch/Streams/ManagedStreamSubject/PostTwitchManagedStreamSubjectDto.cs
// ---------------------------------------------------------------------------------------------------------------------
// Imports
// ---------------------------------------------------------------------------------------------------------------------
namespace NovaLab.Api.Twitch.Streams.ManagedStreamSubject;

// ---------------------------------------------------------------------------------------------------------------------
// Code
// ---------------------------------------------------------------------------------------------------------------------

public record PostTwitchManagedStreamSubjectDto(
    string UserId,
    string SelectionName,
    string ObsSubjectTitle,
    string TwitchSubjectTitle,
    bool AutoSelect = false
);