# src/NovaLab/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "NovaLab API",
    "description": "An ASP.NET Core Web API for managing your streams",
    "version": "v1"
  },
  "paths": {
...
    "/api/twitch/managed-stream-subject": {
      "get": {
        "tags": [
          "TwitchManagedStreamSubject"
        ],
        "operationId": "GetManagedStreamSubjects",
        "parameters": [
          {
            "name": "user-id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TwitchManagedStreamSubjectDtoApiResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TwitchManagedStreamSubjectDtoApiResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TwitchManagedStreamSubjectDtoApiResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TwitchManagedStreamSubject"
        ],
        "operationId": "PostManagedStreamSubjects",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostTwitchManagedStreamSubjectDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostTwitchManagedStreamSubjectDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostTwitchManagedStreamSubjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/twitch/managed-stream-subject/select": {
      "post": {
        "tags": [
          "TwitchManagedStreamSubject"
        ],
        "operationId": "PostSelectManagedStreamSubject",
        "parameters": [
          {
            "name": "user-id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subject-id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResult": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/HttpStatusCode"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": { },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "HttpStatusCode": {
        "enum": [
          100,
          101,
          102,
          103,
          200,
          201,
          202,
          203,
          204,
          205,
          206,
          207,
          208,
          226,
          300,
          301,
          302,
          303,
          304,
          305,
          306,
          307,
          308,
          400,
          401,
          402,
          403,
          404,
          405,
          406,
          407,
          408,
          409,
          410,
          411,
          412,
          413,
          414,
          415,
          416,
          417,
          421,
          422,
          423,
          424,
          426,
          428,
          429,
          431,
          451,
          500,
          501,
          502,
          503,
          504,
          505,
          506,
          507,
          508,
          510,
          511
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "PostTwitchManagedStreamSubjectDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "nullable": true
          },
          "selectionName": {
            "type": "string",
            "nullable": true
          },
          "obsSubjectTitle": {
            "type": "string",
            "nullable": true
          },
          "twitchSubjectTitle": {
            "type": "string",
            "nullable": true
          },
          "autoSelect": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "TwitchManagedStreamSubjectDto": {
        "type": "object",
        "properties": {
          "subjectId": {
            "type": "string",
            "format": "uuid"
          },
          "selectionName": {
            "type": "string",
            "nullable": true
          },
          "obsSubjectTitle": {
            "type": "string",
            "nullable": true
          },
          "twitchSubjectTitle": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TwitchManagedStreamSubjectDtoApiResult": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/HttpStatusCode"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TwitchManagedStreamSubjectDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}